refactor: refactor xml and json statement to extend from StatementNode
