+++ server/common/src/main/java/org/eclipse/lsp/cobol/common/model/tree/JsonParseNode.java
@@ -19,6 +19,7 @@ import lombok.Getter;
 import lombok.ToString;
 import org.eclipse.lsp.cobol.common.model.Locality;
 import org.eclipse.lsp.cobol.common.model.NodeType;
+import org.eclipse.lsp.cobol.common.model.tree.statements.StatementNode;
 import org.eclipse.lsp.cobol.common.model.tree.variable.VariableNameAndLocality;
 
 import java.util.List;
@@ -27,7 +28,7 @@ import java.util.List;
 @ToString(callSuper = true)
 @Getter
 @EqualsAndHashCode(callSuper = true)
-public class JsonParseNode extends Node {
+public class JsonParseNode extends StatementNode {
   private final VariableNameAndLocality identifier1;
   private final VariableNameAndLocality identifier2;
   private final List<VariableNameAndLocality> identifier3;
